requirements.txt
# 🔐 Crypto + Signing
cryptography==42.0.5
pyjwt[crypto]==2.8.0

# 🧪 Validation
pydantic==2.7.1

# 🧬 Environment management
python-dotenv==1.0.1

# ⚡ HTTP client for agents (e.g. Apple ID, APIs)
httpx==0.27.0

# 🛠️ CLI + scripting
argparse==1.4.0

# 🔢 Data Science & Numeric Computing
numpy==1.24.3
pandas==2.0.3

# 📧 Email Services
mailjet-rest==1.3.4

# 🤖 Advanced AI & LangChain
langchain==0.2.6
langchain-google-genai==1.0.6
langgraph==0.1.5

# 📄 PDF Processing
PyPDF2==3.0.1

# 📊 Research & Data Feeds
feedparser==6.0.10

# 🧪 Testing
pytest==8.2.2
pytest-cov==5.0.0

# 🌐 Web API Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# 🤖 AI & Machine Learning
google-generativeai==0.8.1

# 📧 Email & Calendar
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.137.0

# 🌐 Web scraping & HTML parsing
beautifulsoup4==4.12.3
lxml==5.2.2

# 📦 Optional OAuth / Apple ID agents
requests==2.32.3


# Research Agent Requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
arxiv>=2.1.3
PyPDF2>=3.0.1
pypdf>=4.0.1
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
feedparser>=6.0.11
requests>=2.31.0
pandas>=2.1.4
numpy>=1.26.2
typing-extensions>=4.8.0
